Modulegraph: move get_instructions to util.
